requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
--trusted-host pypi.org
--trusted-host files.pythonhosted.org

apscheduler==3.7.0
    # via -r requirements.in
asgiref==3.4.1
    # via
    #   channels
    #   daphne
    #   django
attrs==21.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
authlib==0.15.4
    # via -r requirements.in
autobahn==21.3.1
    # via daphne
automat==20.2.0
    # via twisted
certifi==2021.10.8
    # via requests
cffi==1.14.6
    # via cryptography
channels==3.0.4
    # via -r requirements.in
chardet==4.0.0
    # via requests
click==8.0.1
    # via pip-tools
constantly==15.1.0
    # via twisted
cryptography==35.0.0
    # via
    #   authlib
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.2
    # via channels
django==3.2.8
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
django-cors-headers==3.7.0
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==2.10
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==21.3.0
    # via twisted
pep517==0.11.0
    # via pip-tools
pip-tools==6.3.0
    # via -r requirements.in
psycopg2-binary==2.8.6
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.20
    # via cffi
pyopenssl==21.0.0
    # via twisted
pytz==2021.3
    # via
    #   apscheduler
    #   django
    #   tzlocal
requests==2.25.1
    # via -r requirements.in
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   apscheduler
    #   automat
    #   pyopenssl
    #   service-identity
sqlparse==0.4.2
    # via django
tomli==1.2.1
    # via pep517
twisted[tls]==21.7.0
    # via daphne
txaio==21.2.1
    # via autobahn
typing-extensions==3.10.0.2
    # via twisted
tzlocal==2.1
    # via apscheduler
urllib3==1.26.7
    # via requests
wheel==0.37.0
    # via pip-tools
zope.interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
